Implement tracked resources
This change implements 'tracked' resources for for every resource created in the Radius plane. This means that UCP will observe each operation and maintain its own state for the lifecycle of each resource. The purpose of this is to power multiple other pieces of infrastructure. New in this pull-request it's possible to list the resources (of all types) in a resource group. This is served from the 'tracked' resources maintained by UCP and does not need to query all resource providers or their databases. Future changes will implement cascading deletion for resource groups as well as notifications for resource modifications.